Consolidate annotation exports and restructure SQLite module
