--- pub/orm/lib/schema.dart
@@ -1,5 +1,0 @@
-export 'src/schema/annotations/id.dart';
-export 'src/schema/annotations/map.dart';
-export 'src/schema/annotations/model.dart';
-export 'src/schema/annotations/relation.dart';
-export 'src/schema/annotations/schema.dart';
