+++ pub/orm/lib/sqlite/annotations.dart
@@ -1,0 +1,1 @@
+export '../src/schema/annotations.dart';
