+++ pub/orm/lib/src/schema/annotations.dart
@@ -1,7 +1,44 @@
 import 'package:meta/meta.dart';
+import 'package:meta/meta_meta.dart';
 
-enum ReferenceAction { noAction, restrict, cascade, setNull, setDefault }
+/// Annotation to map a model or field to a database name.
+@immutable
+final class Map {
+  final String name;
 
+  /// Creates a mapping annotation with the given database name.
+  const Map(this.name);
+}
+
+/// Marks a typedef as a model in the schema.
+const model = _Model();
+
+@immutable
+@Target({.typedefType})
+final class _Model {
+  @literal
+  const _Model();
+}
+
+/// Actions for referential integrity on update/delete.
+enum ReferenceAction {
+  /// Do not perform any action.
+  noAction,
+
+  /// Reject the update/delete if dependent rows exist.
+  restrict,
+
+  /// Propagate the update/delete to dependent rows.
+  cascade,
+
+  /// Set the foreign key to NULL.
+  setNull,
+
+  /// Set the foreign key to its default value.
+  setDefault,
+}
+
+/// Annotation describing a relation between models.
 @immutable
 final class Relation {
   /// Defines the name of the relationship. In an m-n-relation,
@@ -26,6 +63,7 @@ final class Relation {
   final ReferenceAction? onUpdate;
 
   @literal
+  /// Creates a relation annotation describing a model relationship.
   const Relation({
     required this.references,
     this.name,
