--- pub/orm/lib/src/schema/annotations/map.dart
@@ -1,8 +1,0 @@
-import 'package:meta/meta.dart';
-
-@immutable
-final class Map {
-  final String name;
-
-  const Map(this.name);
-}
